Change font sizes to decimal type

+++ src/ShapeCrawler/Fonts/IFont.cs
@@ -9,5 +9,5 @@ public interface IFont
     /// <summary>
     ///     Gets or sets font size in points.
     /// </summary>
-    int Size { get; set; }
+    decimal Size { get; set; }
 }
